requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
mysql-connector-python==8.0.22
protobuf==3.14.0
requests==2.25.1
six==1.15.0
urllib3==1.26.2
